@@ -1,10 +1,10 @@
 """Training utilities."""
 
-from typing import Iterator
+from collections.abc import Iterator
 
-from ...tensor_ops.creating import concatenate
+from ...tensor_ops.creating import concat
+from ...tensor_ops.reducing import norm
 from ...tensor_ops.reshaping import flatten
-from ...tensor_ops.transforming import norm
 from ..parameter import Parameter
 
 __all__ = ["clip_grad_norm"]
@@ -26,7 +26,7 @@ def clip_grad_norm(parameters: Iterator[Parameter], max_norm: float) -> float:
         Unclipped gradient norm.
     """
     params = list(parameters)
-    grads = concatenate([flatten(p.grad) for p in params if p.grad])
+    grads = concat([flatten(p.grad) for p in params if p.grad])
     grad_norm = norm(grads).item()
 
     if grad_norm <= max_norm:
