@@ -1,15 +1,16 @@
 """Tensor creation and combination operations."""
 
-from typing import Sequence
+from collections.abc import Sequence
+from typing import Optional
 
-from ..base_tensor import Tensor, _AxisLike, _ShapeLike
-from ..dtypes import Dtype, _DtypeLike, _ScalarLike
-from ..engine import Device, _DeviceLike, get_engine
+from ..backend import Device, select_device
+from ..tensors import AxisLike, ShapeLike, Tensor
+from ..typing import DType, ScalarLike, select_dtype
 
 __all__ = [
     "append",
     "arange",
-    "concatenate",
+    "concat",
     "empty",
     "empty_like",
     "full",
@@ -42,15 +43,16 @@ def append(x: Tensor, values: Tensor, axis: int = -1) -> Tensor:
     Tensor
         Tensor containing appended values.
     """
-    return Tensor(x.engine.append(x.data, values.data, axis=axis))
+    return Tensor(x.device.module.append(x.data, values.data, axis=axis))
 
 
 def arange(
     stop: int | float,
-    start: int | float = 0,
-    step: int | float = 1,
-    dtype: _DtypeLike = Dtype.FLOAT64,
-    device: _DeviceLike = Device.CPU,
+    start: int | float = 0.0,
+    step: int | float = 1.0,
+    *,
+    device: Optional[Device] = None,
+    dtype: Optional[DType] = None,
 ) -> Tensor:
     """Returns a tensor of evenly spaced values using a step size within
     a given interval [start, stop).
@@ -63,21 +65,22 @@ def arange(
         Start value. Defaults to ``0``.
     step : int | float, optional
         Spacing between values. Defaults to ``1``.
+    device : Device, optional
+        The device the tensor is stored on. Defaults to ``None``.
     dtype : DtypeLike, optional
-        Datatype of the tensor data. Defaults to :class:`compyute.float32`.
-    device : DeviceLike, optional
-        The device the tensor is stored on. Defaults to :class:`compyute.cpu`.
+        Datatype of the tensor data. Defaults to ``None``.
 
     Returns
     -------
     Tensor
         Tensor of evenly spaced samples.
     """
-    dtype = Dtype(dtype).value
-    return Tensor(get_engine(device).arange(start, stop, step, dtype=dtype))
+    device = select_device(device)
+    dtype = select_dtype(dtype)
+    return Tensor(device.module.arange(start, stop, step, dtype.type))
 
 
-def concatenate(tensors: Sequence[Tensor], axis: _AxisLike = -1) -> Tensor:
+def concat(tensors: Sequence[Tensor], axis: AxisLike = -1) -> Tensor:
     """Returns a new tensor by joining a sequence of tensors along a given axis.
 
     Parameters
@@ -92,13 +95,15 @@ def concatenate(tensors: Sequence[Tensor], axis: _AxisLike = -1) -> Tensor:
     Tensor
         Concatenated tensor.
     """
-    return Tensor(tensors[0].engine.concatenate([t.data for t in tensors], axis=axis))
+    data = tensors[0].device.module.concatenate([t.data for t in tensors], axis=axis)
+    return Tensor(data)
 
 
 def empty(
-    shape: _ShapeLike,
-    dtype: _DtypeLike = Dtype.FLOAT32,
-    device: _DeviceLike = Device.CPU,
+    shape: ShapeLike,
+    *,
+    device: Optional[Device] = None,
+    dtype: Optional[DType] = None,
 ) -> Tensor:
     """Returns an tensor with uninitialized values.
 
@@ -106,18 +111,19 @@ def empty(
     ----------
     shape : ShapeLike
         Shape of the new tensor.
+    device : Device, optional
+        The device the tensor is stored on. Defaults to ``None``.
     dtype : DtypeLike, optional
-        Datatype of the tensor data. Defaults to :class:`compyute.float32`.
-    device : DeviceLike, optional
-        The device the tensor is stored on. Defaults to :class:`compyute.cpu`.
+        Datatype of the tensor data. Defaults to ``None``.
 
     Returns
     -------
     Tensor
         Tensor with uninitialized values.
     """
-    dtype = Dtype(dtype).value
-    return Tensor(get_engine(device).empty(shape=shape, dtype=dtype))
+    device = select_device(device)
+    dtype = select_dtype(dtype)
+    return Tensor(device.module.empty(shape, dtype.type))
 
 
 def empty_like(x: Tensor) -> Tensor:
@@ -137,10 +143,11 @@ def empty_like(x: Tensor) -> Tensor:
 
 
 def full(
-    shape: _ShapeLike,
-    value: _ScalarLike,
-    dtype: _DtypeLike = Dtype.FLOAT32,
-    device: _DeviceLike = Device.CPU,
+    shape: ShapeLike,
+    value: ScalarLike,
+    *,
+    device: Optional[Device] = None,
+    dtype: Optional[DType] = None,
 ) -> Tensor:
     """Returns a tensor of a given shape with all values being one.
 
@@ -150,21 +157,22 @@ def full(
         Shape of the new tensor.
     value : ScalarLike
         Value to fill the tensor.
+    device : Device, optional
+        The device the tensor is stored on. Defaults to ``None``.
     dtype : DtypeLike, optional
-        Datatype of the tensor data. Defaults to :class:`compyute.float32`.
-    device : DeviceLike, optional
-        The device the tensor is stored on. Defaults to :class:`compyute.cpu`.
+        Datatype of the tensor data. Defaults to ``None``.
 
     Returns
     -------
     Tensor
         Tensor with all values being one.
     """
-    dtype = Dtype(dtype).value
-    return Tensor(get_engine(device).full(shape, value, dtype=dtype))
+    dtype = select_dtype(dtype)
+    device = select_device(device)
+    return Tensor(device.module.full(shape, value, dtype.type))
 
 
-def full_like(x: Tensor, value: _ScalarLike) -> Tensor:
+def full_like(x: Tensor, value: ScalarLike) -> Tensor:
     """Returns a tensor of a given shape with all values being one.
 
     Parameters
@@ -183,7 +191,10 @@ def full_like(x: Tensor, value: _ScalarLike) -> Tensor:
 
 
 def identity(
-    n: int, dtype: _DtypeLike = Dtype.FLOAT32, device: _DeviceLike = Device.CPU
+    n: int,
+    *,
+    device: Optional[Device] = None,
+    dtype: Optional[DType] = None,
 ) -> Tensor:
     """Returns a diagonal tensor of shape ``(n, n)``.
 
@@ -191,26 +202,28 @@ def identity(
     ----------
     n : int
         Size of the new tensor. The shape will be ``(n, n)``.
+    device : Device, optional
+        The device the tensor is stored on. Defaults to ``None``.
     dtype : DtypeLike, optional
-        Datatype of the tensor data. Defaults to :class:`compyute.float32`.
-    device : DeviceLike, optional
-        The device the tensor is stored on. Defaults to :class:`compyute.cpu`.
+        Datatype of the tensor data. Defaults to ``None``.
 
     Returns
     -------
     Tensor
         Diagonal tensor.
     """
-    dtype = Dtype(dtype).value
-    return Tensor(get_engine(device).identity(n, dtype=dtype))
+    dtype = select_dtype(dtype)
+    device = select_device(device)
+    return Tensor(device.module.identity(n, dtype.type))
 
 
 def linspace(
     start: float,
     stop: float,
     num: int,
-    dtype: _DtypeLike = Dtype.FLOAT32,
-    device: _DeviceLike = Device.CPU,
+    *,
+    device: Optional[Device] = None,
+    dtype: Optional[DType] = None,
 ) -> Tensor:
     """Returns a tensor of num evenly spaced values within
     a given interval :math:`[start, stop]`.
@@ -223,24 +236,26 @@ def linspace(
         Stop value.
     num : int
         Number of samples.
+    device : Device, optional
+        The device the tensor is stored on. Defaults to ``None``.
     dtype : DtypeLike, optional
-        Datatype of the tensor data. Defaults to :class:`compyute.float32`.
-    device : DeviceLike, optional
-        The device the tensor is stored on. Defaults to :class:`compyute.cpu`.
+        Datatype of the tensor data. Defaults to ``None``.
 
     Returns
     -------
     Tensor
         Tensor of evenly spaced samples.
     """
-    dtype = Dtype(dtype).value
-    return Tensor(get_engine(device).linspace(start, stop, num, dtype=dtype))
+    dtype = select_dtype(dtype)
+    device = select_device(device)
+    return Tensor(device.module.linspace(start, stop, num, dtype=dtype.type))
 
 
 def ones(
-    shape: _ShapeLike,
-    dtype: _DtypeLike = Dtype.FLOAT32,
-    device: _DeviceLike = Device.CPU,
+    shape: ShapeLike,
+    *,
+    device: Optional[Device] = None,
+    dtype: Optional[DType] = None,
 ) -> Tensor:
     """Returns a tensor of a given shape with all values being one.
 
@@ -248,18 +263,19 @@ def ones(
     ----------
     shape : ShapeLike
         Shape of the new tensor.
+    device : Device, optional
+        The device the tensor is stored on. Defaults to ``None``.
     dtype : DtypeLike, optional
-        Datatype of the tensor data. Defaults to :class:`compyute.float32`.
-    device : DeviceLike, optional
-        The device the tensor is stored on. Defaults to :class:`compyute.cpu`.
+        Datatype of the tensor data. Defaults to ``None``.
 
     Returns
     -------
     Tensor
         Tensor with all values being one.
     """
-    dtype = Dtype(dtype).value
-    return Tensor(get_engine(device).ones(shape, dtype=dtype))
+    dtype = select_dtype(dtype)
+    device = select_device(device)
+    return Tensor(device.module.ones(shape, dtype.type))
 
 
 def ones_like(x: Tensor) -> Tensor:
@@ -297,10 +313,10 @@ def split(x: Tensor, splits: int | Sequence[int], axis: int = -1) -> list[Tensor
     list[Tensor]
         List of tensors containing the split data.
     """
-    return [Tensor(s) for s in x.engine.split(x.data, splits, axis=axis)]
+    return [Tensor(s) for s in x.device.module.split(x.data, splits, axis)]
 
 
-def stack(tensors: Sequence[Tensor], axis: _AxisLike = 0) -> Tensor:
+def stack(tensors: Sequence[Tensor], axis: AxisLike = 0) -> Tensor:
     """Returns a new tensor by stacking a sequence of tensors along a given axis.
 
     Parameters
@@ -315,13 +331,14 @@ def stack(tensors: Sequence[Tensor], axis: _AxisLike = 0) -> Tensor:
     Tensor
         Stacked tensor.
     """
-    return Tensor(tensors[0].engine.stack([t.data for t in tensors], axis=axis))
+    return Tensor(tensors[0].device.module.stack([t.data for t in tensors], axis))
 
 
 def zeros(
-    shape: _ShapeLike,
-    dtype: _DtypeLike = Dtype.FLOAT32,
-    device: _DeviceLike = Device.CPU,
+    shape: ShapeLike,
+    *,
+    device: Optional[Device] = None,
+    dtype: Optional[DType] = None,
 ) -> Tensor:
     """Returns a tensor of a given shape with all values being zero.
 
@@ -329,18 +346,19 @@ def zeros(
     ----------
     shape : ShapeLike
         Shape of the new tensor.
+    device : Device, optional
+        The device the tensor is stored on. Defaults to ``None``.
     dtype : DtypeLike, optional
-        Datatype of the tensor data. Defaults to :class:`compyute.float32`.
-    device : DeviceLike, optional
-        The device the tensor is stored on. Defaults to :class:`compyute.cpu`.
+        Datatype of the tensor data. Defaults to ``None``.
 
     Returns
     -------
     Tensor
         Tensor with all values being zero.
     """
-    dtype = Dtype(dtype).value
-    return Tensor(get_engine(device).zeros(shape, dtype=dtype))
+    dtype = select_dtype(dtype)
+    device = select_device(device)
+    return Tensor(device.module.zeros(shape, dtype.type))
 
 
 def zeros_like(x: Tensor) -> Tensor:
