@@ -1,10 +1,13 @@
 """Neural network functions."""
 
-from .activations import *
-from .convolutions import *
-from .embeddings import *
-from .linear import *
-from .losses import *
-from .metrics import *
-from .normalizatons import *
-from .regularizations import *
+from .activation_funcs import *
+from .convolution_funcs import *
+from .embedding_funcs import *
+from .functions import *
+from .linear_funcs import *
+from .loss_funcs import *
+from .metric_funcs import *
+from .normalization_funcs import *
+from .recurrent_funcs import *
+from .regularization_funcs import *
+from .reshape_funcs import *
