@@ -3,11 +3,11 @@
 from .activations import *
 from .blocks import *
 from .containers import *
-from .convolution import *
-from .embedding import *
+from .convolutions import *
+from .embeddings import *
 from .linear import *
 from .module import *
-from .normalization import *
-from .recurrent import *
-from .regularization import *
-from .reshape import *
+from .normalizations import *
+from .recurrents import *
+from .regularizations import *
+from .reshapes import *
