@@ -2,37 +2,71 @@
 
 from typing import Literal, Optional, TypeAlias
 
-from ...base_tensor import Tensor
-from ..functional.activations import (
-    gelu,
-    leaky_relu,
-    relu,
-    sigmoid,
-    silu,
-    softmax,
-    tanh,
+from ...tensors import Tensor
+from ..functional.activation_funcs import (
+    FastGELUFn,
+    GELUFn,
+    LeakyReLUFn,
+    ReLUFn,
+    SigmoidFn,
+    SiLUFn,
+    SoftmaxFn,
+    TanhFn,
 )
 from .module import Module
 
-__all__ = ["ReLU", "LeakyReLU", "GELU", "Sigmoid", "SiLU", "Softmax", "Tanh"]
+__all__ = [
+    "ReLU",
+    "LeakyReLU",
+    "GELU",
+    "FastGELU",
+    "Sigmoid",
+    "SiLU",
+    "Softmax",
+    "Tanh",
+]
 
 
-class ReLU(Module):
-    r"""Applies the Rectified Linear Unit activation function to an input tensor.
+class GELU(Module):
+    r"""Gaussian Error Linear Unit activation function (using the :math:`tanh` approximation).
 
     .. math::
-        y = \text{max}(0, x)
+        y = 0.5 \cdot x \cdot \left( 1 + \text{tanh} \left( x \sqrt{\frac{2}{\pi}} \cdot (1 + 0.044715 \cdot x^2) \right) \right)
 
     Parameters
     ----------
     label : str, optional
         Module label. Defaults to ``None``. If ``None``, the class name is used.
-
     """
 
+    @Module.register_forward
     def forward(self, x: Tensor) -> Tensor:
-        y, self._backward = relu(x, self._is_training)
-        return y
+        return GELUFn.forward(self.fcache, x)
+
+    @Module.register_backward
+    def backward(self, dy: Tensor) -> Tensor:
+        return GELUFn.backward(self.fcache, dy)
+
+
+class FastGELU(Module):
+    r"""Gaussian Error Linear Unit activation function (using the :math:`sigmoid` approximation).
+
+    .. math::
+        y = x \cdot \sigma{1.702x}
+
+    Parameters
+    ----------
+    label : str, optional
+        Module label. Defaults to ``None``. If ``None``, the class name is used.
+    """
+
+    @Module.register_forward
+    def forward(self, x: Tensor) -> Tensor:
+        return FastGELUFn.forward(self.fcache, x)
+
+    @Module.register_backward
+    def backward(self, dy: Tensor) -> Tensor:
+        return FastGELUFn.backward(self.fcache, dy)
 
 
 class LeakyReLU(Module):
@@ -53,43 +87,35 @@ class LeakyReLU(Module):
         super().__init__(label)
         self.alpha = alpha
 
+    @Module.register_forward
     def forward(self, x: Tensor) -> Tensor:
-        y, self._backward = leaky_relu(x, self.alpha, self._is_training)
-        return y
+        return LeakyReLUFn.forward(self.fcache, x, self.alpha)
+
+    @Module.register_backward
+    def backward(self, dy: Tensor) -> Tensor:
+        return LeakyReLUFn.backward(self.fcache, dy)
 
 
-class GELU(Module):
-    r"""Gaussian Error Linear Unit activation function (using the :math:`tanh` approximation).
+class ReLU(Module):
+    r"""Applies the Rectified Linear Unit activation function to an input tensor.
 
     .. math::
-        y = 0.5 \cdot x \cdot (1 + \text{tanh}(\sqrt{\frac{2}{pi}} \cdot (x + 0.044715 \cdot x^3)))
+        y = \text{max}(0, x)
 
     Parameters
     ----------
     label : str, optional
         Module label. Defaults to ``None``. If ``None``, the class name is used.
+
     """
 
+    @Module.register_forward
     def forward(self, x: Tensor) -> Tensor:
-        y, self._backward = gelu(x, self._is_training)
-        return y
+        return ReLUFn.forward(self.fcache, x)
 
-
-class Tanh(Module):
-    r"""Tanh activation function.
-
-    .. math::
-        y = \text{tanh}(x)
-
-    Parameters
-    ----------
-    label : str, optional
-        Module label. Defaults to ``None``. If ``None``, the class name is used.
-    """
-
-    def forward(self, x: Tensor) -> Tensor:
-        y, self._backward = tanh(x, self._is_training)
-        return y
+    @Module.register_backward
+    def backward(self, dy: Tensor) -> Tensor:
+        return ReLUFn.backward(self.fcache, dy)
 
 
 class Sigmoid(Module):
@@ -104,9 +130,13 @@ class Sigmoid(Module):
         Module label. Defaults to ``None``. If ``None``, the class name is used.
     """
 
+    @Module.register_forward
     def forward(self, x: Tensor) -> Tensor:
-        y, self._backward = sigmoid(x, self._is_training)
-        return y
+        return SigmoidFn.forward(self.fcache, x)
+
+    @Module.register_backward
+    def backward(self, dy: Tensor) -> Tensor:
+        return SigmoidFn.backward(self.fcache, dy)
 
 
 class SiLU(Module):
@@ -121,9 +151,13 @@ class SiLU(Module):
         Module label. Defaults to ``None``. If ``None``, the class name is used.
     """
 
+    @Module.register_forward
     def forward(self, x: Tensor) -> Tensor:
-        y, self._backward = silu(x, self._is_training)
-        return y
+        return SiLUFn.forward(self.fcache, x)
+
+    @Module.register_backward
+    def backward(self, dy: Tensor) -> Tensor:
+        return SiLUFn.backward(self.fcache, dy)
 
 
 class Softmax(Module):
@@ -138,12 +172,37 @@ class Softmax(Module):
         Module label. Defaults to ``None``. If ``None``, the class name is used.
     """
 
+    @Module.register_forward
     def forward(self, x: Tensor) -> Tensor:
-        y, self._backward = softmax(x, self._is_training)
-        return y
+        return SoftmaxFn.forward(self.fcache, x)
+
+    @Module.register_backward
+    def backward(self, dy: Tensor) -> Tensor:
+        return SoftmaxFn.backward(self.fcache, dy)
 
 
-_ActivationLike: TypeAlias = Literal[
+class Tanh(Module):
+    r"""Tanh activation function.
+
+    .. math::
+        y = \text{tanh}(x)
+
+    Parameters
+    ----------
+    label : str, optional
+        Module label. Defaults to ``None``. If ``None``, the class name is used.
+    """
+
+    @Module.register_forward
+    def forward(self, x: Tensor) -> Tensor:
+        return TanhFn.forward(self.fcache, x)
+
+    @Module.register_backward
+    def backward(self, dy: Tensor) -> Tensor:
+        return TanhFn.backward(self.fcache, dy)
+
+
+ActivationLike: TypeAlias = Literal[
     "relu", "leaky_relu", "gelu", "sigmoid", "silu", "tanh"
 ]
 ACTIVATIONS = {
@@ -156,7 +215,7 @@ ACTIVATIONS = {
 }
 
 
-def get_activation(activation: _ActivationLike) -> Module:
+def get_activation(activation: ActivationLike) -> Module:
     """Returns an actiation function."""
 
     if activation not in ACTIVATIONS:
