@@ -1,57 +1,41 @@
 """Neural network parameter initializers."""
 
 import math
-import operator
-from abc import ABC, abstractmethod
-from functools import reduce
+from collections.abc import Callable
+from functools import partial
 from typing import Literal, Optional
 
-from ...base_tensor import Tensor, _ShapeLike
 from ...random.random import normal, uniform
 from ...tensor_ops.creating import ones, zeros
-from ..modules.activations import _ActivationLike
-from ..parameter import Parameter
+from ...tensors import ShapeLike, Tensor
+from ..modules.activations import ActivationLike
 
 __all__ = [
-    "KaimingNormal",
-    "KaimingUniform",
-    "Normal",
-    "Uniform",
-    "XavierNormal",
-    "XavierUniform",
+    "init_kaiming_normal",
+    "init_kaiming_uniform",
+    "init_normal",
+    "init_ones",
+    "init_uniform",
+    "init_xavier_normal",
+    "init_xavier_uniform",
+    "init_zeros",
 ]
 
 
-class Initializer(ABC):
-    """Initializer base class."""
-
-    @abstractmethod
-    def __call__(self, tensor: Tensor | Parameter) -> None:
-        """Fills a tensor with values.
-
-        Parameters
-        ----------
-        tensor : Tensor | Parameter
-            Tensor or parameter to fill with values.
-        """
+def init_ones(*tensors: Tensor) -> None:
+    """Initializes tensors with ones."""
+    for t in tensors:
+        t.data = ones(t.shape, device=t.device, dtype=t.dtype).data
 
 
-class Ones(Initializer):
-    """Initializes a tensor with ones."""
-
-    def __call__(self, tensor: Tensor | Parameter) -> None:
-        tensor.data = ones(tensor.shape, tensor.dtype, tensor.device).data
-
-
-class Zeros(Initializer):
-    """Initializes a tensor with zeros."""
-
-    def __call__(self, tensor: Tensor | Parameter) -> None:
-        tensor.data = zeros(tensor.shape, tensor.dtype, tensor.device).data
+def init_zeros(*tensors: Tensor) -> None:
+    """Initializes tensors with zeros."""
+    for t in tensors:
+        t.data = zeros(t.shape, device=t.device, dtype=t.dtype).data
 
 
-class Normal(Initializer):
-    """Initializes a tensor with values following a uniform distribution.
+def init_normal(*tensors: Tensor, mean: float = 0.0, std: float = 1.0) -> None:
+    """Initializes tensors with values following a uniform distribution.
 
     Parameters
     ----------
@@ -60,19 +44,12 @@ class Normal(Initializer):
     std : float, optional
         Standard deviation of the normal distribution. Defaults to ``1``.
     """
-
-    def __init__(self, mean: float = 0.0, std: float = 1.0) -> None:
-        self.mean = mean
-        self.std = std
-
-    def __call__(self, tensor: Tensor | Parameter) -> None:
-        tensor.data = normal(
-            tensor.shape, self.mean, self.std, tensor.dtype, tensor.device
-        ).data
+    for t in tensors:
+        t.data = normal(t.shape, mean, std, device=t.device, dtype=t.dtype).data
 
 
-class Uniform(Initializer):
-    """Initializes a tensor with values following a uniform distribution.
+def init_uniform(*tensors: Tensor, low: float = -1.0, high: float = 1.0) -> None:
+    """Initializes tensors with values following a uniform distribution.
 
     Parameters
     ----------
@@ -81,18 +58,13 @@ class Uniform(Initializer):
     high : float, optional
         Upper bound for random values. Defaults to ``1``.
     """
-
-    def __init__(self, low: float = -1.0, high: float = 1.0) -> None:
-        self.low = low
-        self.high = high
-
-    def __call__(self, tensor: Tensor | Parameter) -> None:
-        tensor.data = uniform(
-            tensor.shape, self.low, self.high, tensor.dtype, tensor.device
-        ).data
+    for t in tensors:
+        t.data = uniform(t.shape, low, high, device=t.device, dtype=t.dtype).data
 
 
-class KaimingNormal(Initializer):
+def init_kaiming_normal(
+    *tensors: Tensor, activation: Optional[ActivationLike] = None
+) -> None:
     """Normal initializer as described by
     `He et al., 2015 <https://arxiv.org/pdf/1502.01852>`_.
 
@@ -101,17 +73,16 @@ class KaimingNormal(Initializer):
     activation : _ActivationLike
         Activation function to infer the gain value for.
     """
-
-    def __init__(self, activation: Optional[_ActivationLike] = None) -> None:
-        self.gain = get_gain(activation)
-
-    def __call__(self, tensor: Tensor | Parameter) -> None:
-        fan_in = get_fan_in(tensor.shape)
-        std = self.gain / math.sqrt(fan_in)
-        tensor.data = normal(tensor.shape, 0, std, tensor.dtype, tensor.device).data
+    gain = _get_gain(activation)
+    for t in tensors:
+        fan_in = _get_fan_in(t.shape)
+        std = gain / math.sqrt(fan_in)
+        t.data = normal(t.shape, 0, std, device=t.device, dtype=t.dtype).data
 
 
-class KaimingUniform(Initializer):
+def init_kaiming_uniform(
+    *tensors: Tensor, activation: Optional[ActivationLike] = None
+) -> None:
     """Uniform initializer as described by
     `He et al., 2015 <https://arxiv.org/pdf/1502.01852>`_.
 
@@ -120,17 +91,16 @@ class KaimingUniform(Initializer):
     activation : _ActivationLike
         Activation function to infer the gain value for.
     """
-
-    def __init__(self, activation: Optional[_ActivationLike] = None) -> None:
-        self.gain = get_gain(activation)
-
-    def __call__(self, tensor: Tensor | Parameter) -> None:
-        fan_in = get_fan_in(tensor.shape)
-        k = self.gain * math.sqrt(3 / fan_in)
-        tensor.data = uniform(tensor.shape, -k, k, tensor.dtype, tensor.device).data
+    gain = _get_gain(activation)
+    for t in tensors:
+        fan_in = _get_fan_in(t.shape)
+        k = gain * math.sqrt(3 / fan_in)
+        t.data = uniform(t.shape, -k, k, device=t.device, dtype=t.dtype).data
 
 
-class XavierNormal(Initializer):
+def init_xavier_normal(
+    *tensors: Tensor, activation: Optional[ActivationLike] = None
+) -> None:
     """Normal initializer as described by
     `Glorot et al., 2010 <https://proceedings.mlr.press/v9/glorot10a.html>`_.
 
@@ -139,18 +109,17 @@ class XavierNormal(Initializer):
     activation : _ActivationLike
         Activation function to infer the gain value for.
     """
-
-    def __init__(self, activation: Optional[_ActivationLike] = None) -> None:
-        self.gain = get_gain(activation)
-
-    def __call__(self, tensor: Tensor | Parameter) -> None:
-        fan_in = get_fan_in(tensor.shape)
-        fan_out = _get_fan_out(tensor.shape)
-        std = self.gain * math.sqrt(2 / (fan_in + fan_out))
-        tensor.data = normal(tensor.shape, 0, std, tensor.dtype, tensor.device).data
+    gain = _get_gain(activation)
+    for t in tensors:
+        fan_in = _get_fan_in(t.shape)
+        fan_out = _get_fan_out(t.shape)
+        std = gain * math.sqrt(2 / (fan_in + fan_out))
+        t.data = normal(t.shape, 0, std, device=t.device, dtype=t.dtype).data
 
 
-class XavierUniform(Initializer):
+def init_xavier_uniform(
+    *tensors: Tensor, activation: Optional[ActivationLike] = None
+) -> None:
     """Uniform initializer as described by
     `Glorot et al., 2010 <https://proceedings.mlr.press/v9/glorot10a.html>`_.
 
@@ -159,40 +128,34 @@ class XavierUniform(Initializer):
     activation : _ActivationLike
         Activation function to infer the gain value for.
     """
-
-    def __init__(self, activation: Optional[_ActivationLike] = None) -> None:
-        self.gain = get_gain(activation)
-
-    def __call__(self, tensor: Tensor | Parameter) -> None:
-        fan_in = get_fan_in(tensor.shape)
-        fan_out = _get_fan_out(tensor.shape)
-        k = self.gain * math.sqrt(6 / (fan_in + fan_out))
-        tensor.data = uniform(tensor.shape, -k, k, tensor.dtype, tensor.device).data
+    gain = _get_gain(activation)
+    for t in tensors:
+        fan_in = _get_fan_in(t.shape)
+        fan_out = _get_fan_out(t.shape)
+        k = gain * math.sqrt(6 / (fan_in + fan_out))
+        t.data = uniform(t.shape, -k, k, device=t.device, dtype=t.dtype).data
 
 
-_InitializerLike = (
-    Initializer
-    | Literal[
-        "kaiming_normal",
-        "kaiming_uniform",
-        "normal",
-        "uniform",
-        "xavier_normal",
-        "xavier_uniform",
-        "zeros",
-        "ones",
-    ]
-)
+InitializerLike = Literal[
+    "kaiming_normal",
+    "kaiming_uniform",
+    "normal",
+    "uniform",
+    "xavier_normal",
+    "xavier_uniform",
+    "zeros",
+    "ones",
+]
 
-INITIALIZERS = {
-    "kaiming_normal": KaimingNormal,
-    "kaiming_uniform": KaimingUniform,
-    "normal": Normal,
-    "uniform": Uniform,
-    "xavier_normal": XavierNormal,
-    "xavier_uniform": XavierUniform,
-    "zeros": Zeros,
-    "ones": Ones,
+INITIALIZERS: dict[str, Callable[..., None]] = {
+    "kaiming_normal": init_kaiming_normal,
+    "kaiming_uniform": init_kaiming_uniform,
+    "normal": init_normal,
+    "uniform": init_uniform,
+    "xavier_normal": init_xavier_normal,
+    "xavier_uniform": init_xavier_uniform,
+    "zeros": init_zeros,
+    "ones": init_ones,
 }
 
 GAINS = {
@@ -204,31 +167,17 @@ GAINS = {
 }
 
 
-def get_gain(activation: Optional[_ActivationLike] = None) -> float:
-    """Returns the gain value to use for initializing values.
-
-    Parameters
-    ----------
-    activation : _ActivationLike, optional
-        Activation function to infer the gain value for. Defaults to ``None``.
-        If ``None``, a value of ``1`` is returned.
-
-    Returns
-    -------
-    float
-        Gain value.
-    """
+def _get_gain(activation: Optional[ActivationLike] = None) -> float:
+    """Returns the gain value to use for initializing values."""
     if activation is None:
         return 1
     return GAINS.get(activation, 1)
 
 
 def get_initializer(
-    initializer: _InitializerLike, activation: _ActivationLike
-) -> Initializer:
+    initializer: InitializerLike, activation: ActivationLike
+) -> Callable[..., None]:
     """Returns an instance of an initializer."""
-    if isinstance(initializer, Initializer):
-        return initializer
     if initializer not in INITIALIZERS:
         raise ValueError(f"Unknown initializer: {initializer}.")
     if initializer in {
@@ -237,15 +186,15 @@ def get_initializer(
         "kaiming_normal",
         "kaiming_uniform",
     }:
-        return INITIALIZERS[initializer](activation)
-    return INITIALIZERS[initializer]()
+        return partial(INITIALIZERS[initializer], activation=activation)
+    return INITIALIZERS[initializer]
 
 
-def get_fan_in(shape: _ShapeLike) -> int:
+def _get_fan_in(shape: ShapeLike) -> int:
     """Returns the fan-in value for a given shape."""
-    return reduce(operator.mul, (shape[0],) + shape[2:])
+    return math.prod(shape[1:])
 
 
-def _get_fan_out(shape: _ShapeLike) -> int:
+def _get_fan_out(shape: ShapeLike) -> int:
     """Returns the fan-out value for a given shape."""
-    return reduce(operator.mul, (shape[1],) + shape[2:])
+    return math.prod((shape[0],) + shape[2:])
