@@ -2,14 +2,10 @@
 
 from typing import Optional
 
-from ..base_tensor import Tensor, _AxisLike
-from ..dtypes import INT_DTYPES
 from ..random.random import shuffle
 from ..tensor_ops.creating import identity
-from ..tensor_ops.transforming import max as cpmax
-from ..tensor_ops.transforming import mean
-from ..tensor_ops.transforming import min as cpmin
-from ..tensor_ops.transforming import var
+from ..tensors import AxisLike, Tensor
+from ..typing import DType, int64, is_integer
 
 __all__ = ["split_train_val_test", "normalize", "standardize", "one_hot_encode"]
 
@@ -38,8 +34,8 @@ def split_train_val_test(
         Test split.
     """
     x_shuffled = shuffle(x)[0]
-    n1 = int(len(x_shuffled) * (1 - ratio_val - ratio_test))
-    n2 = int(len(x_shuffled) * (1 - ratio_test))
+    n1 = int(len(x_shuffled) * (1.0 - ratio_val - ratio_test))
+    n2 = int(len(x_shuffled) * (1.0 - ratio_test))
     train = x_shuffled[:n1]
     val = x_shuffled[n1:n2]
     test = x_shuffled[n2:]
@@ -48,9 +44,9 @@ def split_train_val_test(
 
 def normalize(
     x: Tensor,
-    axis: Optional[_AxisLike] = None,
-    l_bound: int = 0,
-    u_bound: int = 1,
+    axis: Optional[AxisLike] = None,
+    l_bound: float = 0.0,
+    u_bound: float = 1.0,
 ) -> Tensor:
     """Normalizes a tensor using min-max feature scaling.
 
@@ -61,9 +57,9 @@ def normalize(
     axis : _AxisLike, optional
         Axes over which normalization is applied. Defaults to ``None``.
         If ``None``, the flattended tensor is normalized.
-    l_bound : int, optional
+    l_bound : float, optional
         Lower bound of output values. Defaults to ``0``.
-    u_bound : int, optional
+    u_bound : float, optional
         Upper bound of output values. Defaults to ``1``.
 
     Returns
@@ -72,14 +68,14 @@ def normalize(
         Normalized tensor.
     """
 
-    x_min = cpmin(x, axis=axis)
-    x_max = cpmax(x, axis=axis)
+    x_min = x.min(axis=axis)
+    x_max = x.max(axis=axis)
     return (x - x_min) * (u_bound - l_bound) / (x_max - x_min) + l_bound
 
 
 def standardize(
     x: Tensor,
-    axis: Optional[_AxisLike] = None,
+    axis: Optional[AxisLike] = None,
 ) -> Tensor:
     """Standardizes a tensor to mean 0 and variance 1.
 
@@ -97,10 +93,12 @@ def standardize(
         Standardized tensor with mean ``0`` and variance ``1``.
     """
 
-    return (x - mean(x, axis=axis)) / var(x, axis=axis)
+    return (x - x.mean(axis=axis)) / x.std(axis=axis)
 
 
-def one_hot_encode(x: Tensor, num_classes: int) -> Tensor:
+def one_hot_encode(
+    x: Tensor, num_classes: int, dtype: Optional[DType] = int64
+) -> Tensor:
     """One-hot-encodes a tensor, by adding an additional encoding dimension.
 
     Parameters
@@ -110,6 +108,8 @@ def one_hot_encode(x: Tensor, num_classes: int) -> Tensor:
     num_classes : int
         Number of classes to be considered when encoding.
         Defines axis ``-1`` of the output tensor.
+    dtype : DtypeLike, optional
+        Datatype of the tensor data. Defaults to ``None``.
 
     Returns
     -------
@@ -121,7 +121,7 @@ def one_hot_encode(x: Tensor, num_classes: int) -> Tensor:
     ValueError
         If the tensor dtype is not ``int``.
     """
-    if x.dtype not in INT_DTYPES:
+    if not is_integer(x.dtype):
         raise ValueError(f"Input must be an integer, got '{x.dtype}'.")
 
-    return identity(n=num_classes, dtype=x.dtype, device=x.device)[x]
+    return identity(num_classes, device=x.device, dtype=dtype)[x]
