@@ -11,8 +11,6 @@ from typing import Optional
 import regex
 from tqdm.auto import trange
 
-from ..base_tensor import Tensor, tensor
-
 __all__ = ["CharacterTokenizer", "WordTokenizer", "BPETokenizer"]
 
 WORD_PATTERN = r'([,.:;?_!"()\']|--|\s)'
@@ -21,6 +19,10 @@ BPE_PATTERN = r"""'(?i:[sdmt]|ll|ve|re)|[^\r\n\p{L}\p{N}]?+\p{L}+|\p{N}{1,3}| ?[
 
 class Tokenizer(ABC):
     """Tokenizer base class."""
+
+    oov_token: str
+    vocab: dict
+    ivocab: dict
 
     def __init__(
         self,
@@ -37,7 +39,7 @@ class Tokenizer(ABC):
         """Number of unique tokens."""
         return len(self.vocab)
 
-    def fit(self, text: str, vocab_size: int | None = None) -> None:
+    def fit(self, text: str, vocab_size: Optional[int] = None) -> None:
         """Fits the tokenizer to text.
 
         Parameters
@@ -49,7 +51,7 @@ class Tokenizer(ABC):
         """
 
     @abstractmethod
-    def encode(self, text: str) -> Tensor:
+    def encode(self, text: str) -> list[int]:
         """Encodes text to token ids.
 
         Parameters
@@ -59,18 +61,18 @@ class Tokenizer(ABC):
 
         Returns
         -------
-        Tensor
-            Tensor of token ids.
+        list[int]
+            List of token ids.
         """
 
     @abstractmethod
-    def decode(self, token_ids: Tensor) -> str:
+    def decode(self, token_ids: list[int]) -> str:
         """Decodes token ids to text..
 
         Parameters
         ----------
-        token_ids : Tensor
-            Tensor of integer token ids to be decoded.
+        token_ids : list[int]
+            List of integer token ids to be decoded.
 
         Returns
         -------
@@ -105,12 +107,12 @@ class CharacterTokenizer(Tokenizer):
         ivocab = {token: idx for idx, token in vocab.items()}
         super().__init__(oov_token, vocab, ivocab)
 
-    def encode(self, text: str) -> Tensor:
+    def encode(self, text: str) -> list[int]:
         preprocessed = list(text)
-        return tensor([self.ivocab[s] if s in self.ivocab else 0 for s in preprocessed])
+        return [self.ivocab[s] if s in self.ivocab else 0 for s in preprocessed]
 
-    def decode(self, token_ids: Tensor) -> str:
-        return "".join([self.vocab[i.item()] for i in token_ids])
+    def decode(self, token_ids: list[int]) -> str:
+        return "".join([self.vocab[i] for i in token_ids])
 
 
 class WordTokenizer(Tokenizer):
@@ -128,13 +130,13 @@ class WordTokenizer(Tokenizer):
         self.vocab = dict(enumerate(tokens))
         self.ivocab = {t: i for i, t in self.vocab.items()}
 
-    def encode(self, text: str) -> Tensor:
+    def encode(self, text: str) -> list[int]:
         split = re.split(r'([,.:;?_!"()\']|--|\s)', text)
         split = [s for s in split if s not in [" ", ""]]
-        return tensor([self.ivocab[s] if s in self.ivocab else 0 for s in split])
+        return [self.ivocab[s] if s in self.ivocab else 0 for s in split]
 
-    def decode(self, token_ids: Tensor) -> str:
-        text = " ".join([self.vocab[i.item()] for i in token_ids])
+    def decode(self, token_ids: list[int]) -> str:
+        text = " ".join([self.vocab[i] for i in token_ids])
         text = re.sub(r'\s+([,.:?!"()\'])', r"\1", text)
         return re.sub(r"\n\s", "\n", text)
 
@@ -145,12 +147,16 @@ class BPETokenizer(Tokenizer):
     Mostly follows Andrjey Karpathy's `minbpe <https://github.com/karpathy/minbpe>`_.
     """
 
-    def __init__(self) -> None:
-        super().__init__()
-        self.merges = {}
-        self.pattern = regex.compile(BPE_PATTERN)
+    _merges: dict
+    _pattern: regex.Pattern
 
-    def fit(self, text: str, vocab_size: int = 256) -> None:
+    def __init__(self, oov_token: str = "<|unk|>") -> None:
+        super().__init__(oov_token)
+        self._merges = {}
+        self._pattern = regex.compile(BPE_PATTERN)
+
+    def fit(self, text: str, vocab_size: Optional[int] = None) -> None:
+        vocab_size = vocab_size or 256
         self.vocab = {idx: bytes([idx]) for idx in range(256)}
 
         if vocab_size <= 256:
@@ -159,7 +165,7 @@ class BPETokenizer(Tokenizer):
         n_merges = vocab_size - 256
 
         # split text into chunks according to a regex pattern
-        text_chunks = regex.findall(self.pattern, text)
+        text_chunks = regex.findall(self._pattern, text)
 
         # encode all chunks
         token_ids = [list(chunk.encode("utf-8")) for chunk in text_chunks]
@@ -167,7 +173,7 @@ class BPETokenizer(Tokenizer):
         for i in trange(n_merges, desc="Merges", unit="merges"):
 
             # get counts for bigrams
-            counts = {}
+            counts: dict[tuple[int], int] = {}
             for chunk_ids in token_ids:
                 self._update_counts(chunk_ids, counts)
 
@@ -175,13 +181,13 @@ class BPETokenizer(Tokenizer):
             if len(counts) == 0:
                 print(f"Step {i+1}/{n_merges}. No more possible merges found.")
                 break
-            bigram = max(counts, key=counts.get)
+            bigram: tuple[int, int] = max(counts, key=counts.get)
 
             # replace occurences of bigram with merge id
             idx = 256 + i
             token_ids = [self._merge(chunk_ids, bigram, idx) for chunk_ids in token_ids]
 
-            self.merges[bigram] = idx
+            self._merges[bigram] = idx
             self.vocab[idx] = self.vocab[bigram[0]] + self.vocab[bigram[1]]
 
     def _update_counts(self, token_ids, counts=None):
@@ -216,16 +222,16 @@ class BPETokenizer(Tokenizer):
             counts = self._update_counts(token_ids)
 
             # get bigram that first occured in merges
-            bigram = min(counts, key=lambda p: self.merges.get(p, float("inf")))
-            if bigram not in self.merges:
+            bigram = min(counts, key=lambda p: self._merges.get(p, float("inf")))
+            if bigram not in self._merges:
                 break
 
-            idx = self.merges[bigram]
+            idx = self._merges[bigram]
             token_ids = self._merge(token_ids, bigram, idx)
         return token_ids
 
-    def encode(self, text: str) -> Tensor:
-        text_chunks = regex.findall(self.pattern, text)
+    def encode(self, text: str) -> list[int]:
+        text_chunks = regex.findall(self._pattern, text)
         token_ids = []
 
         for chunk in text_chunks:
@@ -233,13 +239,12 @@ class BPETokenizer(Tokenizer):
             chunk_ids = self._encode_chunk(chunk_bytes)
             token_ids.extend(chunk_ids)
 
-        return tensor(token_ids)
+        return token_ids
 
-    def decode(self, token_ids: Tensor) -> str:
+    def decode(self, token_ids: list[int]) -> str:
         part_bytes = []
 
         for idx in token_ids:
-            idx = idx.item()
             if idx not in self.vocab:
                 raise ValueError(f"invalid token id: {idx}")
             part_bytes.append(self.vocab[idx])
@@ -253,6 +258,6 @@ class BPETokenizer(Tokenizer):
             oov_token=self.oov_token,
             vocab=self.vocab,
             ivocab=self.ivocab,
-            merges=self.merges,
-            pattern=self.pattern,
+            merges=self._merges,
+            pattern=self._pattern,
         )
