@@ -2,13 +2,25 @@
 
 from typing import Optional
 
-from ..base_tensor import ShapeError, Tensor, _AxisLike, tensor
+from ..tensors import AxisLike, ShapeError, Tensor, to_arraylike
+from ..typing import ScalarLike
 
-__all__ = ["argmax", "get_diagonal", "topk", "tril", "triu", "unique"]
+__all__ = [
+    "argmax",
+    "get_diagonal",
+    "max",
+    "maximum",
+    "min",
+    "minimum",
+    "topk",
+    "tril",
+    "triu",
+    "unique",
+]
 
 
 def argmax(
-    x: Tensor, axis: Optional[_AxisLike] = None, keepdims: bool = False
+    x: Tensor, axis: Optional[AxisLike] = None, *, keepdims: bool = False
 ) -> Tensor:
     """Returns the indices of maximum values along a given axis.
 
@@ -16,7 +28,7 @@ def argmax(
     ----------
     x : Tensor
         Input tensor.
-    axis : _AxisLike, optional
+    axis : AxisLike, optional
         Axes, along which the maximum value is located. Defaults to ``None``.
         If ``None`` it is computed over the flattened tensor.
     keepdims : bool, optional
@@ -28,7 +40,7 @@ def argmax(
     Tensor
         Tensor containing indices.
     """
-    return tensor(x.engine.argmax(x.data, axis=axis, keepdims=keepdims))
+    return x.argmax(axis, keepdims=keepdims)
 
 
 def get_diagonal(x: Tensor, d: int = 0) -> Tensor:
@@ -51,10 +63,94 @@ def get_diagonal(x: Tensor, d: int = 0) -> Tensor:
     """
     if x.n_axes < 2:
         raise ShapeError("Input tensor must have at least 2 dimensions.")
-    return Tensor(x.engine.diag(x.data, k=d))
+    return Tensor(x.device.module.diag(x.data, d))
 
 
-def topk(x: Tensor, k: int, axis: _AxisLike = -1) -> tuple[Tensor, Tensor]:
+def max(
+    x: Tensor, axis: Optional[AxisLike] = None, *, keepdims: bool = False
+) -> Tensor:
+    """Computes the maximum of tensor elements over a given axis.
+
+    Parameters
+    ----------
+    x : Tensor
+        Input tensor.
+    axis : AxisLike, optional
+        Axis over which the maximum is computed. Defaults to ``None``.
+        If none it is computed over the flattened tensor.
+    keepdims : bool, optional
+        Whether to keep the tensors dimensions. Defaults to ``False``.
+        if ``False`` the tensor is collapsed along the given axis.
+
+    Returns
+    -------
+    Tensor
+        Tensor containing the maximum of elements.
+    """
+    return x.max(axis, keepdims=keepdims)
+
+
+def maximum(x1: Tensor, x2: Tensor | ScalarLike) -> Tensor:
+    """Computes the element-wise maximum of two tensors or a tensor and a scalar.
+
+    Parameters
+    ----------
+    x1 : Tensor
+        First tensor.
+    x2 : Tensor | ScalarLike
+        Second tensor or scalar.
+
+    Returns
+    -------
+    Tensor
+        Tensor containing the element-wise maximum.
+    """
+    return Tensor(x1.device.module.maximum(x1.data, to_arraylike(x2)))
+
+
+def min(
+    x: Tensor, axis: Optional[AxisLike] = None, *, keepdims: bool = False
+) -> Tensor:
+    """Computes the minimum of tensor elements over a given axis.
+
+    Parameters
+    ----------
+    x : Tensor
+        Input tensor.
+    axis : AxisLike, optional
+        Axis over which the minimum is computed. Defaults to ``None``.
+        If ``None`` it is computed over the flattened tensor.
+    keepdims : bool, optional
+        Whether to keep the tensors dimensions. Defaults to ``False``.
+        if ``False`` the tensor is collapsed along the given axis.
+
+    Returns
+    -------
+    Tensor
+        Tensor containing the minimum of elements.
+    """
+    return x.min(axis, keepdims=keepdims)
+
+
+def minimum(x1: Tensor, x2: Tensor | ScalarLike) -> Tensor:
+    """Computes the element-wise minimum of two tensors or a tensor and a scalar.
+
+    Parameters
+    ----------
+    x1 : Tensor
+        First tensor.
+    x2 : Tensor | ScalarLike
+        Second tensor or scalar.
+
+    Returns
+    -------
+    Tensor
+        Tensor containing the element-wise minimum.
+    """
+    return Tensor(x1.device.module.minimum(x1.data, to_arraylike(x2)))
+
+
+def topk(x: Tensor, k: int, axis: AxisLike = -1) -> tuple[Tensor, Tensor]:
     """Returns the k largest elements along a given axis.
     Implementation by https://hippocampus-garden.com/numpy_topk/.
 
@@ -72,27 +168,27 @@ def topk(x: Tensor, k: int, axis: _AxisLike = -1) -> tuple[Tensor, Tensor]:
     tuple[Tensor, Tensor]
         Tuple containing the top k elements and their indices.
     """
-    ind = x.engine.argpartition(-x.data, k, axis=axis)
-    ind = x.engine.take(ind, x.engine.arange(k), axis=axis)
-    data = x.engine.take_along_axis(-x.data, ind, axis=axis)
+    ind = x.device.module.argpartition(-x.data, k, axis=axis)
+    ind = x.device.module.take(ind, x.device.module.arange(k), axis=axis)
+    data = x.device.module.take_along_axis(-x.data, ind, axis=axis)
 
     # sort within k elements
-    ind_part = x.engine.argsort(data, axis=axis)
-    ind = x.engine.take_along_axis(ind, ind_part, axis=axis)
+    ind_part = x.device.module.argsort(data, axis=axis)
+    ind = x.device.module.take_along_axis(ind, ind_part, axis=axis)
 
-    val = x.engine.take_along_axis(-data, ind_part, axis=axis)
-    return tensor(val), tensor(ind)
+    val = x.device.module.take_along_axis(-data, ind_part, axis=axis)
+    return Tensor(val), Tensor(ind)
 
 
-def tril(x: Tensor, d: int = 0) -> Tensor:
+def tril(x: Tensor, diag_index: int = 0) -> Tensor:
     """Returns the lower triangle of a tensor below the
-    d-th diagonal of the last two dimensions.
+    diagonal of the last two dimensions.
 
     Parameters
     ----------
     x : Tensor
         Input tensor.
-    d : int, optional
+    diag_index : int, optional
         | Index of the diagonal. Defaults to ``0``.
         | ``0``: main diagonal
         | ``> 0``: above main diagonal
@@ -105,18 +201,18 @@ def tril(x: Tensor, d: int = 0) -> Tensor:
     """
     if x.n_axes < 2:
         raise ShapeError("Input tensor must have at least 2 dimensions.")
-    return Tensor(x.engine.tril(x.data, k=d))
+    return Tensor(x.device.module.tril(x.data, diag_index))
 
 
-def triu(x: Tensor, d: int = 0) -> Tensor:
+def triu(x: Tensor, diag_index: int = 0) -> Tensor:
     """Returns the upper triangle of a tensor above the
-    d-th diagonal of the last two dimensions.
+    diagonal of the last two dimensions.
 
     Parameters
     ----------
     x : Tensor
         Input tensor.
-    d : int, optional
+    diag_index : int, optional
         | Index of the diagonal. Defaults to ``0``.
         | ``0``: main diagonal
         | ``> 0``: above main diagonal
@@ -129,7 +225,7 @@ def triu(x: Tensor, d: int = 0) -> Tensor:
     """
     if x.n_axes < 2:
         raise ShapeError("Input tensor must have at least 2 dimensions.")
-    return Tensor(x.engine.triu(x.data, k=d))
+    return Tensor(x.device.module.triu(x.data, diag_index))
 
 
 def unique(x: Tensor) -> Tensor:
@@ -145,4 +241,4 @@ def unique(x: Tensor) -> Tensor:
     Tensor
         Tensor containing unique values.
     """
-    return tensor(x.engine.unique(x.data))
+    return Tensor(x.device.module.unique(x.data))
